+++ test/datatypes.cxx
@@ -836,59 +836,70 @@ static inline void free_2d(void** arr, size_t N) {
 } // namespace MultiDimArrays
 
 MultiDimArrays::DataHolder::DataHolder() {
-    m_short = allocate_2d<short>(5, 7);
-    m_unsigned_short = allocate_2d<unsigned short>(5, 7);
-    m_int = allocate_2d<int>(5, 7);
-    m_unsigned_int = allocate_2d<unsigned int>(5, 7);
-    m_long = allocate_2d<long>(5, 7);
-    m_unsigned_long = allocate_2d<unsigned long>(5, 7);
-    m_long_long = allocate_2d<long long>(5, 7);
-    m_unsigned_long_long = allocate_2d<unsigned long long>(5, 7);
-    m_float = allocate_2d<float>(5, 7);
-    m_double = allocate_2d<double>(5, 7);
+    m_short1              = allocate_2d<short>(5, 7);
+    m_unsigned_short1     = allocate_2d<unsigned short>(5, 7);
+    m_int1                = allocate_2d<int>(5, 7);
+    m_unsigned_int1       = allocate_2d<unsigned int>(5, 7);
+    m_long1               = allocate_2d<long>(5, 7);
+    m_unsigned_long1      = allocate_2d<unsigned long>(5, 7);
+    m_long_long1          = allocate_2d<long long>(5, 7);
+    m_unsigned_long_long1 = allocate_2d<unsigned long long>(5, 7);
+    m_float1              = allocate_2d<float>(5, 7);
+    m_double1             = allocate_2d<double>(5, 7);
 
     for (size_t i = 0; i < 5; ++i) {
         for (size_t j = 0; j < 7; ++j) {
             size_t val = 5*i+j;
-            m_short[i][j]              = (short)val;
-            m_unsigned_short[i][j]     = (unsigned short)val;
-            m_int[i][j]                = (int)val;
-            m_unsigned_int[i][j]       = (unsigned int)val;
-            m_long[i][j]               = (long)val;
-            m_unsigned_long[i][j]      = (unsigned long)val;
-            m_long_long[i][j]          = (long long)val;
-            m_unsigned_long_long[i][j] = (unsigned long long)val;
-            m_float[i][j]              = (float)val;
-            m_double[i][j]             = (double)val;
+            m_short1[i][j]              = (short)val;
+            m_unsigned_short1[i][j]     = (unsigned short)val;
+            m_int1[i][j]                = (int)val;
+            m_unsigned_int1[i][j]       = (unsigned int)val;
+            m_long1[i][j]               = (long)val;
+            m_unsigned_long1[i][j]      = (unsigned long)val;
+            m_long_long1[i][j]          = (long long)val;
+            m_unsigned_long_long1[i][j] = (unsigned long long)val;
+            m_float1[i][j]              = (float)val;
+            m_double1[i][j]             = (double)val;
         }
     }
+
+    m_short2              = nullptr;
+    m_unsigned_short2     = nullptr;
+    m_int2                = nullptr;
+    m_unsigned_int2       = nullptr;
+    m_long2               = nullptr;
+    m_unsigned_long2      = nullptr;
+    m_long_long2          = nullptr;
+    m_unsigned_long_long2 = nullptr;
+    m_float2              = nullptr;
+    m_double2             = nullptr;
 
     for (size_t i = 0; i < 3; ++i) {
         for (size_t j = 0; j < 5; ++j) {
             size_t val = 3*i+j;
-            m_short2[i][j]              = (short)val;
-            m_unsigned_short2[i][j]     = (unsigned short)val;
-            m_int2[i][j]                = (int)val;
-            m_unsigned_int2[i][j]       = (unsigned int)val;
-            m_long2[i][j]               = (long)val;
-            m_unsigned_long2[i][j]      = (unsigned long)val;
-            m_long_long2[i][j]          = (long long)val;
-            m_unsigned_long_long2[i][j] = (unsigned long long)val;
-            m_float2[i][j]              = (float)val;
-            m_double2[i][j]             = (double)val;
+            m_short3[i][j]              = (short)val;
+            m_unsigned_short3[i][j]     = (unsigned short)val;
+            m_int3[i][j]                = (int)val;
+            m_unsigned_int3[i][j]       = (unsigned int)val;
+            m_long3[i][j]               = (long)val;
+            m_unsigned_long3[i][j]      = (unsigned long)val;
+            m_long_long3[i][j]          = (long long)val;
+            m_unsigned_long_long3[i][j] = (unsigned long long)val;
+            m_float3[i][j]              = (float)val;
+            m_double3[i][j]             = (double)val;
         }
     }
 }
 
 MultiDimArrays::DataHolder::~DataHolder() {
-    free_2d((void**)m_short, 5);
-    free_2d((void**)m_unsigned_short, 5);
-    free_2d((void**)m_int, 5);
-    free_2d((void**)m_unsigned_int, 5);
-    free_2d((void**)m_long, 5);
-    free_2d((void**)m_unsigned_long, 5);
-    free_2d((void**)m_long_long, 5);
-    free_2d((void**)m_unsigned_long_long, 5);
-    free_2d((void**)m_float, 5);
-    free_2d((void**)m_double, 5);
+    free_2d((void**)m_short1, 5);
+    free_2d((void**)m_unsigned_short1, 5);
+    free_2d((void**)m_int1, 5);
+    free_2d((void**)m_unsigned_int1, 5);
+    free_2d((void**)m_long1, 5);
+    free_2d((void**)m_unsigned_long1, 5);
+    free_2d((void**)m_long_long1, 5);
+    free_2d((void**)m_unsigned_long_long1, 5);
+    free_2d((void**)m_float1, 5);
+    free_2d((void**)m_double1, 5);
 }